review: changelog the expansion loss, fix stale flag text, rehome shell_escape
- CHANGELOG entry under Fixed. The people this affects are the ones who
  were relying on the container shell to expand a $VAR or a glob out of a
  single argument, and that loss is invisible in the diff, so it says so
  explicitly and names the `bash -lc` form to move to.

- The "You must either provide a --command (-c)" error names a flag that
  does not exist: the command is a trailing positional, and -c is
  unassigned (-C is --config). Reworded to "a command to run", and the
  test asserting on that string follows it.

- shell_escape moves from utils::runs_on to a new utils::shell. A generic
  quoting helper living in the remote-execution module meant a local,
  non-remote path had to reach into runs_on for it. Its four tests move
  with it, and the import is a `use` rather than an inline path now.
