+++ src/utils/mod.rs
@@ -28,6 +28,7 @@ pub mod runs_on;
 pub mod runtime;
 pub mod runtime_extension;
 pub mod scheduler;
+pub mod shell;
 pub mod signing_keys;
 #[cfg(unix)]
 pub mod signing_service;
